iOS build fixes for gateway independence and account controller wiring
- `MixnetConfig` passes `enableNotifications` into `GatewayIndependence` and threads server-family reminder settings from `AppSettings`.
- Adds missing `serverFamilyRemindersEnabled` storage and `setGatewayIndependenceNotifications` through tunnel messaging and macOS gRPC.
- Replaces missing Rust `VpnError.AccountStoreBusy` with Swift-only `VPNErrorReason.accountStoreBusy` for the tunnel-active guard.
- Splits iOS account-controller session code into `CredentialsManager+iOSAccountController.swift` so SwiftLint stays under the type body error limit.
- Excludes `**/build/**` from local SwiftLint so SourcePackages are not scanned.
