+++ nym-vpn-apple/Services/Sources/Services/AppSettings/AppSettings.swift
@@ -87,6 +87,11 @@ import ConnectionTypes
         didSet { isAdBlockerEnabledPublisher = isAdBlockerEnabled}
     }
 
+    @AppStorage(AppSettingKey.serverFamilyReminders.rawValue)
+    public var serverFamilyRemindersEnabled = true {
+        didSet { serverFamilyRemindersEnabledPublisher = serverFamilyRemindersEnabled }
+    }
+
 #if os(macOS)
     @AppStorage(AppSettingKey.statistics.rawValue)
     public var isStatisticsEnabled = true
@@ -175,6 +180,7 @@ import ConnectionTypes
     @Published public var isMixnetTuningEnabledPublisher: Bool
     @Published public var isAdBlockerEnabledPublisher: Bool
     @Published public var isPassphraseStoredPublisher: Bool
+    @Published public var serverFamilyRemindersEnabledPublisher: Bool
 
     // Init ensures the *Publisher mirrors stored values* on launch.
     private init() {
@@ -189,6 +195,7 @@ import ConnectionTypes
         self.isMixnetTuningEnabledPublisher = false
         isAdBlockerEnabledPublisher = false
         isPassphraseStoredPublisher = false
+        serverFamilyRemindersEnabledPublisher = true
 
         self.isErrorReportingOnPublisher = self.isErrorReportingOn
         self.isCredentialImportedPublisher = self.isCredentialImported
@@ -201,6 +208,7 @@ import ConnectionTypes
         self.isMixnetTuningEnabledPublisher = self.isMixnetTuningEnabled
         self.isAdBlockerEnabledPublisher = self.isAdBlockerEnabled
         self.isPassphraseStoredPublisher = self.isPassphraseStored
+        self.serverFamilyRemindersEnabledPublisher = self.serverFamilyRemindersEnabled
     }
 
     public func resetUserDefaults() {
@@ -255,6 +263,7 @@ public enum AppSettingKey: String {
     case accountSummaryCache
     case accountSummaryLastFetchedAt
     case oneClickDisplayMode
+    case serverFamilyReminders
 }
 
 extension Array: @retroactive RawRepresentable where Element: Codable {
